// src/core/types/components.ts
export type ComponentType =
    | 'animation'
    | 'boxcollider'
    | 'camerafollow'
    | 'camerashake'
    | 'damageradius'
    | 'deadbodyondeath'
    | 'entitydestination'
    | 'entityeffect'
    | 'entityfollow'
    | 'health'
    | 'highlight'
    | 'lifetime'
    | 'lightemit'
    | 'meleeattack'
    | 'particle'
    | 'particleemit'
    | 'playercontrol'
    | 'projectile'
    | 'rangedattackemitter'
    | 'rigidbody'
    | 'script'
    | 'shadow'
    | 'slowtime'
    | 'sound'
    | 'sprite'
    | 'spritestate'
    | 'teleport'
    | 'textlabel'
    | 'transform';